+++ src/Parser/NoHttpRequestParser.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 namespace Yansongda\Pay\Parser;
 
 use Psr\Http\Message\ResponseInterface;
+use Yansongda\Pay\Contract\PackerInterface;
 use Yansongda\Pay\Contract\ParserInterface;
 
 class NoHttpRequestParser implements ParserInterface
 {
-    public function parse(?ResponseInterface $response): ?ResponseInterface
+    public function parse(PackerInterface $packer, ?ResponseInterface $response): ?ResponseInterface
     {
         return $response;
     }