+++ src/Service/UnipayServiceProvider.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Yansongda\Pay\Service;
 
-use Yansongda\Pay\Contract\ParserInterface;
 use Yansongda\Pay\Contract\ServiceProviderInterface;
-use Yansongda\Pay\Parser\CollectionParser;
 use Yansongda\Pay\Pay;
 use Yansongda\Pay\Provider\Unipay;
 
@@ -19,7 +17,6 @@ class UnipayServiceProvider implements ServiceProviderInterface
     {
         $service = new Unipay();
 
-        Pay::set(ParserInterface::class, CollectionParser::class);
         Pay::set(Unipay::class, $service);
         Pay::set('unipay', $service);
     }