+++ tests/Parser/ArrayParserTest.php
@@ -5,18 +5,28 @@ namespace Yansongda\Pay\Tests\Parser;
 use GuzzleHttp\Psr7\Response;
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Pay\Exception\InvalidResponseException;
+use Yansongda\Pay\Packer\JsonPacker;
+use Yansongda\Pay\Packer\QueryPacker;
 use Yansongda\Pay\Parser\ArrayParser;
 use Yansongda\Pay\Tests\TestCase;
 
 class ArrayParserTest extends TestCase
 {
+    protected ArrayParser $parser;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->parser = new ArrayParser();
+    }
+
     public function testResponseNull()
     {
         self::expectException(InvalidResponseException::class);
         self::expectExceptionCode(Exception::RESPONSE_NONE);
 
-        $parser = new ArrayParser();
-        $parser->parse(null);
+        $this->parser->parse(new JsonPacker(), null);
     }
 
     public function testWrongFormat()
@@ -26,16 +36,14 @@ class ArrayParserTest extends TestCase
 
         $response = new Response(200, [], '{"name": "yansongda"}a');
 
-        $parser = new ArrayParser();
-        $parser->parse($response);
+        $this->parser->parse(new JsonPacker(), $response);
     }
 
     public function testNormal()
     {
         $response = new Response(200, [], '{"name": "yansongda"}');
 
-        $parser = new ArrayParser();
-        $result = $parser->parse($response);
+        $result = $this->parser->parse(new JsonPacker(), $response);
 
         self::assertEquals(['name' => 'yansongda'], $result);
     }
@@ -46,8 +54,7 @@ class ArrayParserTest extends TestCase
 
         $response->getBody()->read(2);
 
-        $parser = new ArrayParser();
-        $result = $parser->parse($response);
+        $result = $this->parser->parse(new JsonPacker(), $response);
 
         self::assertEquals(['name' => 'yansongda'], $result);
     }
@@ -56,8 +63,7 @@ class ArrayParserTest extends TestCase
     {
         $response = new Response(200, [], 'name=yansongda&age=29');
 
-        $parser = new ArrayParser();
-        $result = $parser->parse($response);
+        $result = $this->parser->parse(new QueryPacker(), $response);
 
         self::assertEqualsCanonicalizing(['name' => 'yansongda', 'age' => '29'], $result);
     }
@@ -68,8 +74,7 @@ class ArrayParserTest extends TestCase
 
         $response = new Response(200, [], json_encode(['h5_url' => $url]));
 
-        $parser = new ArrayParser();
-        $result = $parser->parse($response);
+        $result = $this->parser->parse(new JsonPacker(), $response);
 
         self::assertEquals('https://yansongda.cn?name=yansongda&age=29', $result['h5_url']);
     }